Implementing the edit, add and remove product

+++ backend/controllers/ProductController.js
@@ -61,7 +61,7 @@ export const updateProduct = async (req, res) => {
 
     let fileName = "";
     if(req.files === null) {
-        fileName = Product.image;
+        fileName = product.image;
     } else {
         const file = req.files.file;
         const fileSize = file.data.length;
@@ -72,9 +72,12 @@ export const updateProduct = async (req, res) => {
         if(!allowedType.includes(ext.toLowerCase())) return res.status(422).json({msg:"Invalid Images"});
         if(fileSize > 5000000) return res.status(422).json({msg: "Image must be less than 5 MB"});
 
+        // Hanya hapus file lama jika ada file baru
         const filePath = `./public/images/${product.image}`;
-        fs.unlinkSync(filePath)
-
+        if (fs.existsSync(filePath)) {
+            fs.unlinkSync(filePath);
+        }
+    
         file.mv(`./public/images/${fileName}`, (err) => {
             if(err) return res.status(500).json({msg: err.message});
         });
